develop employee list modal
