+++ src/components/salary_calculator/basic_info_form.tsx
@@ -3,85 +3,10 @@ import useOuterClick from '@/lib/hooks/use_outer_click';
 import { FaChevronDown } from 'react-icons/fa6';
 import { FiSearch } from 'react-icons/fi';
 import { PiUserFill } from 'react-icons/pi';
-import { RxCross2 } from 'react-icons/rx';
 import { MONTHS } from '@/constants/month';
 import NumericInput from '@/components/numeric_input/numeric_input';
+import EmployeeListModal from '@/components/salary_calculator/employee_list_modal';
 import { useCalculatorCtx } from '@/contexts/calculator_context';
-import { dummyEmployeeForCalc } from '@/interfaces/employees';
-
-const EmployeeListModal: React.FC<{
-  modalVisibleHandler: () => void;
-}> = ({ modalVisibleHandler }) => {
-  // Info: (20250711 - Julian) Get data from API
-  const employeesList = dummyEmployeeForCalc;
-
-  const { changeEmployeeName, changeEmployeeNumber, setBaseSalary } = useCalculatorCtx();
-
-  const displayedEmployeesList = employeesList.map((employee) => {
-    // Info: (20250711 - Julian) 填入員工姓名、編號和本薪
-    const handleClick = () => {
-      changeEmployeeName(employee.name);
-      changeEmployeeNumber(employee.number ?? '');
-      setBaseSalary(employee.baseSalary);
-      modalVisibleHandler();
-    };
-
-    return (
-      <div
-        onClick={handleClick}
-        className="group flex items-center bg-surface-neutral-surface-lv2 px-24px py-12px hover:cursor-pointer hover:bg-surface-brand-primary-soft"
-      >
-        <div className="flex flex-1 items-center gap-8px">
-          <PiUserFill
-            size={16}
-            className="text-text-neutral-tertiary group-hover:text-text-neutral-primary"
-          />
-          <p className="font-medium text-text-neutral-secondary group-hover:text-text-neutral-primary">
-            {employee.name}
-          </p>
-        </div>
-        <div className="flex items-center gap-8px text-sm font-medium">
-          <p className="text-text-neutral-tertiary group-hover:text-text-neutral-primary">#</p>
-          <p className="text-text-neutral-secondary group-hover:text-text-neutral-primary">
-            {employee.number}
-          </p>
-        </div>
-      </div>
-    );
-  });
-
-  return (
-    <div className="fixed inset-0 z-70 flex items-center justify-center bg-black/50 font-barlow">
-      <div className="relative flex flex-col rounded-sm bg-surface-neutral-surface-lv2 md:w-440px">
-        {/* Info: (20250711 - Julian) Modal Header */}
-        <div className="relative flex items-start justify-center px-40px py-16px">
-          <h2 className="text-lg font-bold text-card-text-primary">Employee List</h2>
-          <button type="button" onClick={modalVisibleHandler} className="absolute right-20px">
-            <RxCross2 scale={24} />
-          </button>
-        </div>
-        {/* Info: (20250711 - Julian) Modal Body */}
-        <div className="flex flex-col gap-24px px-40px py-24px">
-          {/* Info: (20250711 - Julian) Search bar */}
-          <div className="flex items-center rounded-sm border border-input-stroke-input">
-            <div className="px-12px py-10px text-icon-surface-single-color-primary">
-              <FiSearch size={16} />
-            </div>
-            <input
-              type="text"
-              placeholder="Search for employee"
-              className="flex-1 bg-transparent px-12px py-10px placeholder:text-input-text-input-placeholder"
-            />
-          </div>
-          {/* Info: (20250711 - Julian) Employee list content */}
-          <div className="flex max-h-500px flex-col divide-y divide-stroke-neutral-quaternary overflow-y-auto pb-30px">
-            {displayedEmployeesList}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 const BasicInfoForm: React.FC = () => {
   const yearOptions = ['2025', '2024', '2023'];
